Refactor Issues Table to a component
