+++ src/routes/projects/[id]/issues/+page.svelte
@@ -3,26 +3,13 @@
   import type { IProject, IIssue } from '../../../../services/sitevisor-types';
   import HeaderProject from '$lib/components/HeaderProject.svelte';
   import AddIssueDialog from '$lib/components/AddIssueDialog.svelte';
-  import { formatDate } from '$lib/utils/helpers';
+  import IssuesTable from '$lib/components/IssuesTable.svelte';
   export let data: PageData;
-  import { get } from 'svelte/store';
-	import { objectTypesStore, statusOptionsStore } from '../../../../stores';
 
   let project: IProject = data.project;
   let issues: IIssue[] = data.issues;
   let projectId: string = project.id.toString()
   let addIssueDialogVisible: boolean = false;
-
-  const statusOptions = get(statusOptionsStore);
-  const objectTypes = get(objectTypesStore);
-
-  $: formattedIssues = issues.map(issue => ({
-    ...issue,
-    status: statusOptions.get(issue.status),
-    object_type: objectTypes.get(issue.object_type),
-    created_at: formatDate(issue.created_at),
-    updated_at: formatDate(issue.updated_at)
-  }));
 
   function openAddIssueDialog() {
     addIssueDialogVisible = true;
@@ -44,39 +31,7 @@
 <HeaderProject projectid={projectId}/>
 <div class="container mx-auto">
   <div class="container mx-auto overflow-y-scroll h-[calc(100vh-170px)] p-5 pt-0">
-    <table class="table w-full table-zebra table-pin-rows">
-      <thead>
-        <tr>
-          <th>Title</th>
-          <th>Status</th>
-          <th>Type</th>
-          <th>Created by</th>
-          <th>Assigned to</th>
-          <th>Created</th>
-          <th>Updated</th>
-          <th></th>
-        </tr>
-      </thead>
-      <tbody>
-        {#each formattedIssues as issue}
-          <tr>
-            <td>{issue.title}</td>
-            <td>{issue.status}</td>
-            <td>{issue.object_type}</td>
-            <td>{issue.creator}</td>
-            <td>{issue.assignee ? issue.assignee : 'Unassigned'}</td>
-            <td>{issue.created_at}</td>
-            <td>{issue.updated_at}</td>
-            <td>
-              <div class="flex justify-end gap-3">
-                <a class="btn btn-xs" href="/projects/{projectId}/issues/{issue.id}">Details</a>
-              </div>
-            </td>
-          </tr>
-        {/each}
-      </tbody>
-    </table>
-
+    <IssuesTable issues={issues} />
   </div>
   <div class="flex justify-end mt-8">
     <button class="btn btn-primary" on:click={openAddIssueDialog}>Add New Issue</button>
